refactor(controller): split library-db.ts into a barrel over library-db/
2433 lines across fourteen banner-delimited sections that were already
conceptually separate — the banners are now the module boundaries. Same
barrel-over-parts shape as the settings.ts split, and every consumer
already imported it as a namespace, so nothing at a call site changes.

  handle.ts       the open handle + shared constants (the no-cycle seam)
  types.ts        record shapes; TrackRow is the raw SQLite row
  rows.ts         row -> record mapping and the JSON column parsers
  stats.ts        library-wide counts, briefly cached
  schema.ts       migrations, versioned by PRAGMA user_version
  legacy.ts       one-shot moods.json -> SQLite import
  lifecycle.ts    open / close / backup / restore / reset
  meta.ts         embedding-model metadata
  tracks.ts       per-track reads and writes
  vectors.ts      sqlite-vec KNN + the sound-map projection
  audio-moods.ts  zero-shot moods scored from CLAP audio vectors
  queries.ts      mood- and tag-keyed reads, genre centroids
  browse.ts       the admin browse filter + Observatory rows
  plays.ts        play history

As with settings/store.ts, the mutable module state needed a seam: db,
currentEmbeddingDim and dbNonce can't be assigned through an import, so
handle.ts owns them behind getDb()/getEmbeddingDim()/getDbNonce() and a
single setHandle() writer that only lifecycle.ts calls. stats.ts sorts
before lifecycle.ts because close() invalidates the stats cache — that
was the one back-edge in the section order.

No behaviour change. Controller lint clean, all 52 test files pass, and
a smoke run covers open → write → read → filter → close → reopen →
persistence → the embedding-dim mismatch guard.
